Move "base" channels.
