+++ homeassistant/components/zha/core/channels/measurement.py
@@ -8,7 +8,6 @@ import logging
 
 import zigpy.zcl.clusters.measurement as measurement
 
-from . import AttributeListeningChannel
 from .. import registries
 from ..const import (
     REPORT_CONFIG_DEFAULT,
@@ -16,6 +15,7 @@ from ..const import (
     REPORT_CONFIG_MAX_INT,
     REPORT_CONFIG_MIN_INT,
 )
+from .base import AttributeListeningChannel
 
 _LOGGER = logging.getLogger(__name__)
