+++ homeassistant/components/zha/core/channels/protocol.py
@@ -9,7 +9,7 @@ import logging
 import zigpy.zcl.clusters.protocol as protocol
 
 from .. import registries
-from ..channels import ZigbeeChannel
+from .base import ZigbeeChannel
 
 _LOGGER = logging.getLogger(__name__)
